Crate apollo

Apollo

A library for generative musical composition and synthesis. Each generated song is an instance of the Apollo struct which is initialised with the following data:

A Config

A ConfigPreset

A seed value for the random number generator

A sample rate

An example invocation that outputs 2 seconds of samples to the console:

use serde_json;
use apollo::{
    Apollo,
    Config,
};
 
// Create basic configuration from JSON
let config: Config = serde_json::from_str(r#"{
    "seed": 1,
    "presets": [{
        "id": "default",
        "rhythm": {
            "bpm": 60,
            "beat": 4,
            "unit": [4],
            "weight": "shorter"
        },
        "scale": ["major"],
        "tracks": [{
            "num": 1,
            "phrase": {
                "length": 2,
                "note": {
                    "frequency": 220.0,
                    "adsr": {
                        "amplitude": 1.0,
                        "attack": 0.4,
                        "release": 0.2
                    }
                }
            },
            "synth": {
                "num": 4,
                "offset": 0.0,
                "signal": ["sine", "saw"],
                "filter": [{
                    "type": "four_pole",
                    "frequency": 300.0,
                    "resonance": 0.5
                }]
            }
        }]
    }],
    "scales": [{
        "name": "C Major",
        "tags": ["major"],
        "notes": [261.63, 293.66, 329.63, 349.23, 392.00, 440.00, 493.88]
    }]
}"#)?;
let preset = String::from("default");
let sample_rate = 44100;
let length = 2 * sample_rate; // 2 seconds
let seed = 1;
 
// Create sample iterator using "default" configuration preset
let apollo = Apollo::new(config, preset, seed, sample_rate as f32)?.into_iter();
 
// Output 2 seconds worth of clock and sample values to console
for (clock, value) in (0..length).zip(apollo) {
    println!("{} {}", clock, value);
}

Structs

Apollo
An iterator that produces audio samples

Config
The configuration for song generation

ConfigADSR
An attack, decay, sustain, release envelope configuration

ConfigChoice
A list of values from which one can be randomly selected

ConfigNote
A musical note configuration

ConfigPhrase
A musical phrase configuration consisting of a series of notes in one or more bars

ConfigPreset
A song preset configuration

ConfigRhythm
A musical rhythm configuration

ConfigScale
A musical scale configuration

ConfigSynth
A synthesizer configuration

ConfigTrack
A track configuration consisting of a musical phrase and instrument

Enums

ConfigError
A configuration error

ConfigRange
A range of values from min to max that a value can be randomly generated within or, alternatively, a single fixed value that will always be used

ConfigRhythmWeight
A note length weighting configuration for this rhythm

ConfigSynthFilter
A synthesizer filter type configuration

ConfigSynthSignal
A synthesizer signal type configuration